refactor: split json schema interface rows
